+++ core/src/com/angian/libgdx/pipes/Water.java
@@ -1,6 +1,5 @@
 package com.angian.libgdx.pipes;
 
-import com.badlogic.gdx.graphics.g2d.Batch;
 import com.badlogic.gdx.math.GridPoint2;
 import com.badlogic.gdx.scenes.scene2d.Stage;
 
@@ -8,13 +7,13 @@ import java.util.ArrayList;
 import java.util.List;
 
 public class Water extends BaseActor {
-    private static final float TILE_CAPACITY = 100.0f;
+    private static final float PIPE_CAPACITY = 100.0f;
 
     private final float speed;
     private final GameBoard board;
 
     private final List<PathItem> path;
-    private float tileOccupation;
+    private float pipeOccupation;
     private boolean stopped;
 
 
@@ -23,13 +22,13 @@ public class Water extends BaseActor {
         speed = sp;
         board = b;
 
-        tileOccupation = TILE_CAPACITY / 2;  // water starts from the center of source tile
+        pipeOccupation = PIPE_CAPACITY / 2;  // water starts from the center of source tile
         stopped = false;
 
         path = new ArrayList<>();
 
         GridPoint2 sourcePos = board.getSourcePos();
-        path.add(new PathItem(sourcePos, board.getTile(sourcePos).anyValidDirection()));
+        path.add(new PathItem(sourcePos, board.getPipe(sourcePos).anyValidDirection()));
     }
 
     public boolean isStopped() {
@@ -45,8 +44,8 @@ public class Water extends BaseActor {
         if (stopped)
             return;
 
-        tileOccupation += (speed * dt);
-        if (tileOccupation >= TILE_CAPACITY) {
+        pipeOccupation += (speed * dt);
+        if (pipeOccupation >= PIPE_CAPACITY) {
             //check next tile
             PathItem lastStep = path.get(path.size() - 1);
             PathItem nextStep = board.followPipe(lastStep.pos, lastStep.outDir);
@@ -55,15 +54,26 @@ public class Water extends BaseActor {
             if (nextStep != null) {
                 System.out.println("Added new step to water path: " + nextStep);
                 path.add(nextStep);
-                tileOccupation -= TILE_CAPACITY;
+                pipeOccupation -= PIPE_CAPACITY;
             } else {
                 System.out.println("!! Game over; total distance: " + path.size());
                 stopped = true;
             }
         }
+
+        Direction fromDir = null;
+        for (int i=0; i < path.size(); i++) {
+            PathItem step = path.get(i);
+
+            Pipe pipe = board.getPipe(step.pos);
+            Direction toDir = step.outDir;
+            if (i < path.size() - 1)
+                pipe.setWaterLevel(1.0f, fromDir, toDir);
+            else
+                pipe.setWaterLevel(pipeOccupation / PIPE_CAPACITY, fromDir, toDir);
+
+            fromDir = toDir.flip();
+        }
     }
 
-    public void draw(Batch batch, float parentAlpha) {
-        //TODO
-    }
 }