+++ core/src/com/angian/libgdx/pipes/WaterTimer.java
@@ -1,6 +1,5 @@
 package com.angian.libgdx.pipes;
 
-import com.badlogic.gdx.graphics.Color;
 import com.badlogic.gdx.graphics.g2d.Batch;
 import com.badlogic.gdx.graphics.glutils.ShapeRenderer;
 import com.badlogic.gdx.math.Rectangle;